# kv/logstore: avoid heap allocations around non-blocking sync waiter callback